feat: added bottom navigation bar, removed from main selection screen
