+++ .metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 5464c5bac742001448fe4fc0597be939379f88ea
+  channel: stable
+
+project_type: app
